+++ src/CoreTests/ClientInputQueueTests.cs
@@ -18,9 +18,9 @@ public sealed class ClientInputQueueTests
     [InlineData(-1L)]
     [InlineData(1L)]
     [InlineData(42L)]
-    [InlineData(long.MaxValue)]
-    [InlineData(long.MinValue)]
-    public void Basic(long id)
+    [InlineData(int.MaxValue)]
+    [InlineData(int.MinValue)]
+    public void Basic(int id)
     {
         ClientInputQueue<MockStructure> queue = new(1, new DefaultClientInputPredictor<MockStructure>(), (_, _, _) => {});
         
@@ -68,18 +68,18 @@ public sealed class ClientInputQueueTests
 
     sealed class InputAuthCapturer
     {
-        public long? IdCapture = null;
+        public int? IdCapture = null;
         public long? FrameCapture = null;
         public TimeSpan? DifferenceCapture = null;
 
-        public void Capture(long id, long frame, TimeSpan difference)
+        public void Capture(int id, long frame, TimeSpan difference)
         {
             IdCapture = id;
             FrameCapture = frame;
             DifferenceCapture = difference;
         }
 
-        public void AssertValid(long id, long frame, TimeSpan min, TimeSpan max)
+        public void AssertValid(int id, long frame, TimeSpan min, TimeSpan max)
         {
             Assert.Equal(id, IdCapture);
             Assert.Equal(frame, FrameCapture);
@@ -98,9 +98,9 @@ public sealed class ClientInputQueueTests
     [InlineData(-1L, 42)]
     [InlineData(1L, 1)]
     [InlineData(42L, 42)]
-    [InlineData(long.MaxValue, 1)]
-    [InlineData(long.MinValue, 42)]
-    public void InputAuthorLate(long id, int tps)
+    [InlineData(int.MaxValue, 1)]
+    [InlineData(int.MinValue, 42)]
+    public void InputAuthorLate(int id, int tps)
     {
         InputAuthCapturer capturer = new();
         ClientInputQueue<MockStructure> queue = new(tps, new DefaultClientInputPredictor<MockStructure>(), capturer.Capture);
@@ -124,9 +124,9 @@ public sealed class ClientInputQueueTests
     [InlineData(-1L, 42)]
     [InlineData(1L, 1)]
     [InlineData(42L, 42)]
-    [InlineData(long.MaxValue, 1)]
-    [InlineData(long.MinValue, 42)]
-    public void InputAuthorInTime(long id, int tps)
+    [InlineData(int.MaxValue, 1)]
+    [InlineData(int.MinValue, 42)]
+    public void InputAuthorInTime(int id, int tps)
     {
         InputAuthCapturer capturer = new();
         ClientInputQueue<MockStructure> queue = new(tps, new DefaultClientInputPredictor<MockStructure>(), capturer.Capture);
